Separate schema, handler, and input elements
Separates the schema creation, the onSubmit & onSubmitInvalid, and the input element logic from the respective entry type.
